+++ tests/xlfunctions_vs_excel/sqrt_test.py
@@ -1,6 +1,6 @@
 from .. import testing
 
-from xlfunctions import xlerrors
+from xlcalculator.xlfunctions import xlerrors
 
 
 class SqrtTest(testing.FunctionalTestCase):